ToyBox/requirements.txt
altgraph==0.17.2
annoy==1.17.0
boto3==1.19.3
botocore==1.22.3
bpy-build==2.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
cloudpickle==2.0.0
cmake==3.22.1
cmake-generators==1.0.9
cycler==0.11.0
distro==1.6.0
docker==5.0.3
fonttools==4.28.5
future-fstrings==1.2.0
gitdb==4.0.9
GitPython==3.1.24
gym==0.21.0
idna==3.3
ikpy==3.1
install==1.3.5
jmespath==0.10.0
joblib==1.1.0
keyboard==0.13.5
kiwisolver==1.3.2
llvmlite==0.37.0
matplotlib==3.5.1
mpmath==1.2.1
nose==1.3.7
numba==0.54.1
numpy==1.20.3
opencv-python==4.5.4.60
overrides==6.1.0
packaging==21.0
pacmap==0.5.3
pandas==1.3.5
Pillow==8.4.0
psutil==5.8.0
pydub==0.25.1
pyelftools==0.27
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyparsing==3.0.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
s3transfer==0.5.0
scikit-learn==1.0.1
scipy==1.7.3
screeninfo==0.8.1
seaborn==0.11.2
six==1.16.0
smmap==5.0.0
svn==1.0.1
sympy==1.9
tdw==1.10.9.0
threadpoolctl==3.0.0
tqdm==4.62.3
typing-utils==0.1.0
typing_extensions==4.0.1
urllib3==1.26.7
websocket-client==1.2.3
